# crypto/cipher: optimize safeXORBytes