Optimize safeXORBytes based on this conversation: https://github.com/golang/go/issues/35381

Before:

BenchmarkXORBytes/8Bytes-12             202142712                5.947 ns/op    1345.12 MB/s
BenchmarkXORBytes/128Bytes-12           19122530                63.56 ns/op     2013.72 MB/s
BenchmarkXORBytes/2048Bytes-12           1309760               908.1 ns/op      2255.28 MB/s
BenchmarkXORBytes/32768Bytes-12            72750             14673 ns/op        2233.20 MB/s
PASS
ok      crypto/cipher   6.437s


After:

BenchmarkXORBytes/8Bytes-12             189982386                5.441 ns/op    1470.39 MB/s
BenchmarkXORBytes/128Bytes-12           42866022                28.41 ns/op     4505.66 MB/s
BenchmarkXORBytes/2048Bytes-12           3013195               406.8 ns/op      5034.92 MB/s
BenchmarkXORBytes/32768Bytes-12           186090              6396 ns/op        5123.55 MB/s
PASS
ok      crypto/cipher   5.808