@tbeeck 2024-09-17
Apologies for abandoning some time ago--as this function was removed from the package with the change to export the XOR implementation, I will close this review